Code cleanup.
I'm not best pleased with the layout of the code, so I've cleaned up a
significant portion of it. I've also begun to refactor the interface.
Rather than looking for a file, DNSimple() now expects to be passed the
username and password when created. It's your job to work out how you
want to do that.