Add shell of missing API methods.
The original version of this library only contained methods for a couple
of the APIs available. I've added shells for the remaining parts. This
likely does not represent the final interface, for two reasons. Firstly,
I've not included any required variables that would contain information
for use in POST requests. Second, implementing the full set of
functionality in this file makes the file very large and difficult to
navigate. I'll likely refactor this stuff out at some stage.